fix(sync): align bloom filter FP rate types and add deserialization validation
- Change SyncProtocol::BloomFilter::false_positive_rate from f64 to f32
  to match the f32 used in DeltaIdBloomFilter::new(), BloomFilterRequest,
  and DEFAULT_BLOOM_FP_RATE

- Add is_valid() method to DeltaIdBloomFilter to check structural
  consistency after deserialization, preventing index-out-of-bounds
  panics from malicious peers crafting filters with num_bits exceeding
  bits.len() * 8

- Add is_valid() method to BloomFilterRequest that delegates to the
  contained filter's validation

- Add tests for the validation methods

+++ crates/node/primitives/src/sync/protocol.rs
@@ -92,7 +92,7 @@ pub enum SyncProtocol {
         /// **Note**: Validation of bounds is performed when constructing the actual
         /// bloom filter, not at protocol negotiation time. Invalid values will cause
         /// filter construction to fail gracefully.
-        false_positive_rate: f64,
+        false_positive_rate: f32,
     },
 
     /// Subtree prefetch for deep localized changes.
